Worked on #2 created item views for different tabs
Wrote Layout Resources for Listings/Courses
Wrote Fragments for Courses/Listings.
Wrote Adapters for Listings/Courses
Populated all three with some sample Data.
Modified the PagerAdapter so that the Tabs are in order.

Still need to work out how we want to store images. Using local
resources for now, but I added a "image_url" Parameter to the book class
constructor, as that is what it will eventually be taking.

That's most of it.